git-svn-id: http://mytardis.googlecode.com/svn/branches/single-search-box-shauno@1516 1bbb43ac-ffe0-11de-8c0c-f1c8ac664a8e

+++ tardis/tardis_portal/forms.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 #
-# Copyright (c) 2010, Monash e-Research Centre
+# Copyright (c) 2010-2011, Monash e-Research Centre
 #   (Monash University, Australia)
-# Copyright (c) 2010, VeRSI Consortium
+# Copyright (c) 2010-2011, VeRSI Consortium
 #   (Victorian eResearch Strategic Initiative, Australia)
 # All rights reserved.
 # Redistribution and use in source and binary forms, with or without
@@ -59,7 +59,7 @@ from registration.models import RegistrationProfile
 
 from tardis.tardis_portal import models
 from tardis.tardis_portal.fields import MultiValueCommaSeparatedField
-from tardis.tardis_portal.widgets import CommaSeparatedInput, Span
+from tardis.tardis_portal.widgets import CommaSeparatedInput, Span, TextInput
 from tardis.tardis_portal.models import UserProfile, UserAuthentication
 from tardis.tardis_portal.auth.localdb_auth \
     import auth_key as locabdb_auth_key
@@ -273,10 +273,12 @@ class MXDatafileSearchForm(DatafileSearchForm):
         required=False, label='Max Resolution Limit')
     xrayWavelengthFrom = forms.IntegerField(
         required=False, label='X-ray Wavelength From',
-        widget=forms.TextInput(attrs={'size': '4'}))
+        widget=forms.TextInput(attrs={'size': '4'})
+        )
     xrayWavelengthTo = forms.IntegerField(
         required=False, label='X-ray Wavelength To',
-        widget=forms.TextInput(attrs={'size': '4'}))
+        widget=forms.TextInput(attrs={'size': '4'})
+        )
 
 
 def createLinkedUserAuthenticationForm(authMethods):
@@ -323,7 +325,7 @@ class RegisterExperimentForm(forms.Form):
     xmldata = forms.FileField()
     experiment_owner = forms.CharField(max_length=400, required=False)
     originid = forms.CharField(max_length=400, required=False)
-
+    from_url = forms.CharField(max_length=400, required=False)
 
 class Author_Experiment(forms.ModelForm):
 
@@ -352,21 +354,26 @@ class FullExperimentModel(UserDict):
             ae.experiment = ae.experiment
             ae.save()
         for ds in self.data['datasets']:
-            ds.experiment = ds.experiment
-            ds.save()
+            if not ds.immutable:
+                ds.experiment = ds.experiment
+                ds.save()
         for ds_f in self.data['dataset_files']:
-            ds_f.dataset = ds_f.dataset
-            ds_f.save()
+            if not ds.immutable:
+                ds_f.dataset = ds_f.dataset
+                ds_f.save()
 
         # XXX because saving the form can be now done without
         # commit=False this won't be called during the creation
         # of new experiments.
         if hasattr(self.data['datasets'], 'deleted_forms'):
             for dataset in self.data['datasets'].deleted_forms:
-                dataset.instance.delete()
+                if not dataset.instance.immutable:
+                    dataset.instance.delete()
+
         if hasattr(self.data['dataset_files'], 'deleted_forms'):
             for dataset in self.data['dataset_files'].deleted_forms:
-                dataset.instance.delete()
+                if not dataset.instance.immutable:
+                    dataset.instance.delete()
 
 
 class DataFileFormSet(BaseInlineFormSet):
@@ -380,7 +387,7 @@ class DataFileFormSet(BaseInlineFormSet):
         super(DataFileFormSet, self).__init__(**kwargs)
 
     def save_new(self, form, commit=True):
-        # this is a local file so correct the missing details
+        #this is a local file so correct the missing details
         datafile = super(DataFileFormSet, self).save_new(form, commit=False)
 
         filepath = form.cleaned_data['filename']
@@ -440,22 +447,35 @@ class ExperimentForm(forms.ModelForm):
                                              empty_permitted=False)
 
         def custom_field_cb(field):
+            #if field.name == 'filename':
+            #    return field.formfield(required=False)
             if field.name == 'url':
-                return field.formfield(required=False)
-            elif field.name == 'filename':
-                return field.formfield(widget=Span)
+                return field.formfield(widget=Span, required=False)
+            else:
+                return field.formfield()
+
+        def custom_dataset_field_cb(field):
+            if field.name == 'description':
+                return field.formfield(
+                    widget=TextInput(attrs={'size': '80'}))
             else:
                 return field.formfield()
 
         # initialise formsets
-        dataset_formset = inlineformset_factory(models.Experiment,
-                                                models.Dataset,
-                                                extra=extra, can_delete=True)
-        datafile_formset = inlineformset_factory(models.Dataset,
-                                         models.Dataset_File,
-                                         formset=DataFileFormSet,
-                                         formfield_callback=custom_field_cb,
-                                         extra=0, can_delete=True)
+        if instance == None or instance.dataset_set.count() == 0:
+            extra = 1
+        dataset_formset = inlineformset_factory(
+            models.Experiment,
+            models.Dataset,
+            formfield_callback=custom_dataset_field_cb,
+            extra=extra, can_delete=True)
+
+        datafile_formset = inlineformset_factory(
+            models.Dataset,
+            models.Dataset_File,
+            formset=DataFileFormSet,
+            formfield_callback=custom_field_cb,
+            extra=0, can_delete=True)
 
         # fix up experiment form
         post_authors = self._parse_authors(data)
@@ -475,9 +495,16 @@ class ExperimentForm(forms.ModelForm):
                                         instance=instance,
                                         prefix="dataset")
         for i, df in enumerate(self.datasets.forms):
+            if 'immutable' in df.initial:
+                if df.initial['immutable']:
+                    df.fields['description'].widget.attrs['readonly'] = True
+                    df.fields['description'].editable = False
+                    df.fields['immutable'].editable = False
+                    df.fields['immutable'].widget.attrs['readonly'] = True
+
             self.dataset_files[i] = datafile_formset(data=data,
                                          instance=df.instance,
-                                         prefix="dataset-%s-datafile" % i)
+                                         prefix="ds-%s-datafile" % i)
 
     def _parse_authors(self, data=None):
         """
@@ -552,23 +579,32 @@ class ExperimentForm(forms.ModelForm):
             o_ae = ae.save(commit=commit)
             author_experiments.append(o_ae)
         for key, dataset in enumerate(self.datasets.forms):
-            # XXX for some random reason the link between the instance needs
-            # to be reinitialised
-            dataset.instance.experiment = experiment
-            o_dataset = dataset.save(commit)
-            datasets.append(o_dataset)
-            # save any datafiles if the data set has any
-            if self.dataset_files[key]:
-                o_df = self.dataset_files[key].save(commit)
-                dataset_files += o_df
+            if dataset not in self.datasets.deleted_forms:
+                # XXX for some random reason the link between
+                # the instance needs
+                # to be reinitialised
+                dataset.instance.experiment = experiment
+                o_dataset = dataset.save(commit)
+                datasets.append(o_dataset)
+                # save any datafiles if the data set has any
+                mutable = True
+                if 'immutable' in dataset.initial:
+                    if dataset.initial['immutable']:
+                        mutable = False
+
+                if self.dataset_files[key] and mutable:
+                    o_df = self.dataset_files[key].save(commit)
+                    dataset_files += o_df
 
         if hasattr(self.datasets, 'deleted_forms'):
             for ds in self.datasets.deleted_forms:
-                ds.instance.delete()
+                if not ds.instance.immutable:
+                    ds.instance.delete()
 
         if hasattr(self.dataset_files, 'deleted_forms'):
             for df in self.dataset_files.deleted_forms:
-                df.instance.delete()
+                if not ds.instance.immutable:
+                    df.instance.delete()
 
         return FullExperimentModel({'experiment': experiment,
                                     'author_experiments': author_experiments,
@@ -702,39 +738,6 @@ def createSearchExperimentForm():
 
     for schema, parameterNames in parameterNameGroups.items():
         formutilFields[schema] = []
-
-        for parameterName in parameterNames:
-            if parameterName.data_type == ParameterName.NUMERIC:
-                if parameterName.comparison_type \
-                    == ParameterName.RANGE_COMPARISON:
-                    fields[parameterName.name + 'From'] = \
-                        forms.DecimalField(label=parameterName.full_name
-                            + ' From', required=False)
-                    fields[parameterName.name + 'To'] = \
-                        forms.DecimalField(label=parameterName.full_name
-                            + ' To', required=False)
-                    formutilFields[schema].append(parameterName.name + 'From')
-                    formutilFields[schema].append(parameterName.name + 'To')
-                else:
-                    # note that we'll also ignore the choices text box entry
-                    # even if it's filled if the parameter is of numeric type
-                    # TODO: decide if we are to raise an exception if
-                    #       parameterName.choices is not empty
-                    fields[parameterName.name] = \
-                        forms.DecimalField(label=parameterName.full_name,
-                            required=False)
-                    formutilFields[schema].append(parameterName.name)
-            else:  # parameter is a string
-                if parameterName.choices != '':
-                    fields[parameterName.name] = \
-                        forms.CharField(label=parameterName.full_name,
-                        widget=forms.Select(choices=__getParameterChoices(
-                        parameterName.choices)), required=False)
-                else:
-                    fields[parameterName.name] = \
-                        forms.CharField(label=parameterName.full_name,
-                        max_length=255, required=False)
-                formutilFields[schema].append(parameterName.name)
 
         for parameterName in parameterNames:
             if parameterName.data_type == ParameterName.NUMERIC:
@@ -961,19 +964,19 @@ def create_datafile_add_form(
             if parameter_name.units:
                 units = " (" + parameter_name.units + ")"
 
-                # if not valid, spit back as exact
-                if parameter_name.isNumeric():
-                    fields[key] = \
-                        forms.DecimalField(label=parameter_name.full_name + units,
-                                           required=False,
-                                           initial=value,
-                                           )
-                else:
-                    fields[key] = \
-                        forms.CharField(label=parameter_name.full_name + units,
-                                        max_length=255, required=False,
-                                        initial=value,
-                                        )
+            # if not valid, spit back as exact
+            if parameter_name.isNumeric():
+                fields[key] = \
+                    forms.DecimalField(label=parameter_name.full_name + units,
+                                       required=False,
+                                       initial=value,
+                                       )
+            else:
+                fields[key] = \
+                    forms.CharField(label=parameter_name.full_name + units,
+                                    max_length=255, required=False,
+                                    initial=value,
+                                    )
 
         return type('DynamicForm', (forms.BaseForm, ), {'base_fields': fields})
